requirements.txt
Flask
Flask-Login
Flask-Script
bson
Flask-PyMongo
Flask-Mobility
Flask-User
Flask-MongoEngine
WTForms
Jinja2
dnspython
gunicorn
pandas
rauth
Werkzeug
argparse
itsdangerous
Flask-Mail
email_validator


goose3
newspaper3k
pytimeextractor
inquirer
spacy
geopy
googlesearch-python
Selenium
webdriver-manager
deep_translator